move setPointPosition to packedpath.py
