+++ src/fontra/core/glyphchanges.py
@@ -1,12 +1,11 @@
 from .changes import baseChangeFunctions
-from .packedpath import deleteContour, insertContour, deletePoint, insertPoint
-
-
-def setPointPosition(path, pointIndex, x, y):
-    coords = path["coordinates"]
-    i = pointIndex * 2
-    coords[i] = x
-    coords[i + 1] = y
+from .packedpath import (
+    deleteContour,
+    deletePoint,
+    insertContour,
+    insertPoint,
+    setPointPosition,
+)
 
 
 glyphChangeFunctions = {
